requirements.txt
openai
pyyaml
cryptography